Finalisation des fonctionnalités
* Scénarios appliquable et supprimable à la simulation
* Nouveau scénario de modification des fréquences d'émission des
messages
* Mise à jour des commentaires
* Modification de l'envoi TCP - ajout d'un bouton Stop et mise en
threading du traitement
* Ajout d'une icone au démarrage (chargement de l'application)
* Ajout d'un menu Aide avec une page A propos et un lien vers la
documentation
* Correction de divers bug
* Ajout d'une vue de prévisualisation
* Ajout dans le menu Menu d'un bouton Nouveau permettant de relancer
l'application
* Ajout d'une demande de confirmation à la fermeture de l'application

+++ AisSimulator/src/main/java/model/aismessages/AisStream.java
@@ -7,6 +7,9 @@ import java.io.IOException;
 import java.io.OutputStreamWriter;
 import java.util.ArrayList;
 
+/**
+ * Modèle correspondant à un flux de messages AIS
+ */
 public class AisStream {
 
 	private ArrayList<String> aisMessages;
@@ -27,10 +30,20 @@ public class AisStream {
 		return aisMessages.size();
 	}
 	
-	public int getTimeStamp(String msg) {
-		return Integer.parseInt(msg.substring(4, msg.indexOf("*")));
+	/**
+	 * Retourne le timeStamp d'un message passé en paramètre
+	 * @param msg le message
+	 * @return long en millisecond
+	 */
+	public long getTimeStamp(String msg) {
+		return Long.parseLong(msg.substring(4, msg.indexOf("*")));
 	}
 	
+	/**
+	 * Créé un object File à partir du flux AIS
+	 * @return File
+	 * @throws IOException
+	 */
 	public File export() throws IOException {
 		File f = new File("aisMessage.log");
 		FileOutputStream fos = new FileOutputStream(f);
